split views.py into several submodules
