+++ annotations/views/__init__.py
@@ -1,0 +1,2 @@
+from annotations.views import aws, data, main, network, project, quadruple, relationtemplate, \
+              rest, search, text, user
